added user into resp obj

+++ backend/utils.py
@@ -4,42 +4,51 @@ from rest_framework.permissions import IsAuthenticated
 
 from .models import Note
 from .serializers import NoteSerializer
-
-@permission_classes([IsAuthenticated])
+ 
 def getNotesList(request):
     user = request.user
-    #returns list of all notes from db, need to serialize into json form
-    notes = Note.objects.all().order_by('-updated')
-    serializer = NoteSerializer(notes, many=True)    #returned as an object
-    return Response(serializer.data)
+    if user.is_authenticated:
+        notes = Note.objects.filter(user=user).order_by('-updated')
+        serializer = NoteSerializer(notes, many=True)
+        return Response(serializer.data)
+    return Response("Unauthorized", status=401)
 
 def getNoteDetail(request, id):
-    #returns list of all notes from db, need to serialize into json form
-    notes = Note.objects.get(id=id)
-    serializer = NoteSerializer(notes, many=False)    #returned as an object, serialize single object
-    return Response(serializer.data)
+    user = request.user
+    if user.is_authenticated:
+        note = Note.objects.filter(user=user, id=id).first()
+        if note:
+            serializer = NoteSerializer(note)
+            return Response(serializer.data)
+        return Response("Note not found", status=404)
+    return Response("Unauthorized", status=401)
 
 def createNote(request):
     data = request.data
-    note = Note.objects.create(
-        body=data
-    )
-    serializer = NoteSerializer(note, many=False)
-    return Response(serializer.data)
+    user = request.user
+    note = Note.objects.create(body=data.get('body'), user=user)
+    serializer = NoteSerializer(note)
+    return Response(serializer.data, status=201)
 
 def updateNote(request, id):
-    #return obj, already in json
-    data = request.data
-    note = Note.objects.get(id=id)
-    #get instance of the note that needs to be updated with new info
-    serializer = NoteSerializer(instance=note, data=data)
-
-    if serializer.is_valid():
-        serializer.save()
-    return Response(serializer.data)
+    user = request.user
+    if user.is_authenticated:
+        note = Note.objects.filter(user=user, id=id).first()
+        if note:
+            serializer = NoteSerializer(instance=note, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            return Response(serializer.errors, status=400)
+        return Response("Note not found", status=404)
+    return Response("Unauthorized", status=401)
 
 def deleteNote(request, id):
-    note = Note.objects.get(id=id)
-    #removes object from database
-    note.delete()
-    return Response('Note was deleted')
+    user = request.user
+    if user.is_authenticated:
+        note = Note.objects.filter(user=user, id=id).first()
+        if note:
+            note.delete()
+            return Response('Note was deleted')
+        return Response("Note not found", status=404)
+    return Response("Unauthorized", status=401)
